# api_glewe/api_glewe/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": " OpenApi Documentation",
        "description": " Swagger OpenApi description",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/offer/search-course-by-category": {
            "post": {
                "tags": [
                    "search-course-by-category"
                ],
                "operationId": "70cd2d20d46db2571a5fc7b2324f90c1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "category_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "listing course in terms of category"
                    }
                }
            }
        },
        "/api/offer/search-course": {
            "post": {
                "tags": [
                    "search-course-by-name"
                ],
                "operationId": "e80906e39817551bb8fa79d0405ff2fc",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "salade"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "listing course in terms of name"
                    }
                }
            }
        },
        "/api/offer/course-detail": {
            "post": {
                "tags": [
                    "detail-course"
                ],
                "operationId": "dd0f2359e3f4e6d91009531b31bc4f38",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "course_id": {
                                        "type": "integer",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "detail course"
                    }
                }
            }
        },
...
        "/api/offer/user-validated-course": {
            "post": {
                "tags": [
                    "finish-course"
                ],
                "operationId": "d4bd31a28c5ec2cd654402a8e39ade3a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "course_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "user_id": {
                                        "type": "string",
                                        "example": "99487892-f338-4740-88e7-b3377eafe173"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "I finished this course"
                    }
                }
            }
        },
...
    }
}